+++ types.js
@@ -19,7 +19,6 @@ module.exports = {
         min   : -Infinity,   // number.min Minimum number value
         max   : Infinity,    // number.max Maximum number value
         check : function( number ){ // number.check check number type and size
-            if(number == parseInt(number)) number = parseInt(number);
 
             return typeof number === 'number'
                     && number >= this.min
@@ -32,7 +31,6 @@ module.exports = {
         min   : -Infinity,   // number.min Minimum number value
         max   : Infinity,    // number.max Maximum number value
         check : function( number ){ // number.check check number type and size
-            if(number == parseFloat(number)) number = parseFloat(number);
 
             return typeof number === 'number'
                     && number >= this.min